gwpy.io.gwf: moved frame reading utils to io.gwf
