+++ gwpy/tests/test_io.py
@@ -25,7 +25,7 @@ import tempfile
 from common import skip_missing_import
 from compat import (unittest, mock)
 
-from gwpy.io import datafind
+from gwpy.io import (datafind, gwf)
 from gwpy.io.cache import (Cache, CacheEntry, cache_segments)
 from gwpy.segments import (Segment, SegmentList)
 
@@ -147,48 +147,53 @@ class CacheIoTestCase(unittest.TestCase):
         self.assertEquals(sl, segs)
 
 
-# -- gwpy.io.datafind ---------------------------------------------------------
+# -- gwpy.io.gwf --------------------------------------------------------------
 
 def mock_call(*args, **kwargs):
     raise OSError("")
 
 
-class DataFindIoTestCase(unittest.TestCase):
+class GwfIoTestCase(unittest.TestCase):
+
     @skip_missing_import('lalframe')
     def test_iter_channel_names(self):
         # maybe need something better?
         from types import GeneratorType
-        names = datafind.iter_channel_names(TEST_GWF_FILE)
+        names = gwf.iter_channel_names(TEST_GWF_FILE)
         self.assertIsInstance(names, GeneratorType)
         self.assertSequenceEqual(list(names), TEST_CHANNELS)
         with mock.patch('gwpy.utils.shell.call', mock_call):
-            names = datafind.iter_channel_names(TEST_GWF_FILE)
+            names = gwf.iter_channel_names(TEST_GWF_FILE)
             self.assertIsInstance(names, GeneratorType)
             self.assertSequenceEqual(list(names), TEST_CHANNELS)
 
     @skip_missing_import('lalframe')
     def test_get_channel_names(self):
-        self.assertListEqual(datafind.get_channel_names(TEST_GWF_FILE),
+        self.assertListEqual(gwf.get_channel_names(TEST_GWF_FILE),
                              TEST_CHANNELS)
 
     @skip_missing_import('lalframe')
     def test_num_channels(self):
-        self.assertEqual(datafind.num_channels(TEST_GWF_FILE), 3)
+        self.assertEqual(gwf.num_channels(TEST_GWF_FILE), 3)
 
     @skip_missing_import('lalframe')
     def test_get_channel_type(self):
-        self.assertEqual(datafind.get_channel_type(
+        self.assertEqual(gwf.get_channel_type(
             'L1:LDAS-STRAIN', TEST_GWF_FILE), 'proc')
-        self.assertRaises(ValueError, datafind.get_channel_type,
+        self.assertRaises(ValueError, gwf.get_channel_type,
                           'X1:NOT-IN_FRAME', TEST_GWF_FILE)
 
     @skip_missing_import('lalframe')
     def test_channel_in_frame(self):
         self.assertTrue(
-            datafind.channel_in_frame('L1:LDAS-STRAIN', TEST_GWF_FILE))
+            gwf.channel_in_frame('L1:LDAS-STRAIN', TEST_GWF_FILE))
         self.assertFalse(
-            datafind.channel_in_frame('X1:NOT-IN_FRAME', TEST_GWF_FILE))
+            gwf.channel_in_frame('X1:NOT-IN_FRAME', TEST_GWF_FILE))
 
+
+# -- gwpy.io.datafind ---------------------------------------------------------
+
+class DataFindIoTestCase(unittest.TestCase):
     def test_on_tape(self):
         self.assertFalse(datafind.on_tape(TEST_GWF_FILE))
         self.assertFalse(datafind.on_tape(
